modularize me
