+++ src/tqdm_dlbar.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+
+from tqdm import tqdm
+
+
+# +-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-+
+# | TQDM WGET WITH PROGRESS BAR                                                                         |
+# +-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-+
+class DLProgressBarTQDM(tqdm):
+    def update_to(self, b=1, bsize=1, tsize=None) -> None:
+        if tsize is not None:
+            self.total = tsize
+        return self.update(b * bsize - self.n)  # also sets self.n = b * bsize
